moving cleanup code to where it's used
